requirements.txt
matplotlib>=3.2.0
numpy>=1.16.0
Pillow>=8.4.0
torch>=1.10.2
torchvision>=0.11.3
tqdm>=4.59.0